+++ src/routes/list/api/getTerms.ts
@@ -1,4 +1,4 @@
-import { AppDb } from 'db';
+import { AppDb } from 'services/db';
 import { Term } from 'types/Term';
 import { computeSkip, Pagination } from 'utils/computeSkip';
